Afegeixo script per realitzar backups de MySQL mitjançant l'eina de Percona XtraBackup

+++ centos70-install-docker.sh
@@ -6,27 +6,29 @@
 yum update -y
 
 #Afegim el repositori de Docker
- tee /etc/yum.repos.d/docker.repo <<-'EOF'
- [dockerrepo]
- name=Docker Repository
- baseurl=https://yum.dockerproject.org/repo/main/centos/7/
- enabled=1
- gpgcheck=1
- gpgkey=https://yum.dockerproject.org/gpg
+tee /etc/yum.repos.d/docker.repo <<-'EOF'
+[dockerrepo]
+name=Docker Repository
+baseurl=https://yum.dockerproject.org/repo/main/centos/7/
+enabled=1
+gpgcheck=1
+gpgkey=https://yum.dockerproject.org/gpg
 EOF
 
+#Si podem afegir el repositori realitzem la instal·lació
+if [ $? -eq 0 ] then
+	# Instal·lem el paquet de docker
+	yum install -y docker-engine
 
-# Instal·lem el paquet de docker
-yum install -y docker-engine
+	# execució del script d'instal·lació (això és el mateix que els dos passos anteriors)
+	#curl -fsSL https://get.docker.com/ | sh
 
-# execució del script d'instal·lació (això és el mateix que els dos passos anteriors)
-#curl -fsSL https://get.docker.com/ | sh
+	# Activem el servei
+	systemctl enable docker.service
 
-# Activem el servei
-systemctl enable docker.service
+	# Iniciem el servei Docker
+	systemctl start docker
 
-# Iniciem el servei Docker
-systemctl start docker
-
-# Verifiquem la instl·lació mitjançant l'execució d'una container de prova
-docker run --rm hello-world
+	# Verifiquem la instl·lació mitjançant l'execució d'una container de prova
+	docker run --rm hello-world
+fi
